Added users and fixed some issues

+++ db/migrate/20141229172007_add_money_to_shift.rb
@@ -1,5 +1,5 @@
 class AddMoneyToShift < ActiveRecord::Migration
   def change
-    add_column :shifts, :money, :double
+    add_column :shifts, :money, :float
   end
 end
