+++ db/migrate/20141231143449_add_closing_cash_to_shift.rb
@@ -1,7 +1,7 @@
 class AddClosingCashToShift < ActiveRecord::Migration
   def change
     remove_column :shifts, :money
-    add_column :shifts, :closing_cash, :double
-    add_column :shifts, :opening_cash, :double
+    add_column :shifts, :closing_cash, :float
+    add_column :shifts, :opening_cash, :float
   end
 end
